requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
mysql-connector-python==8.1.0
nest-asyncio==1.5.7
networkx==3.0
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.28.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.20
stack-data==0.6.2
sympy==1.11.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.159
urllib3==1.26.13
wcwidth==0.2.6
